+++ pom.xml
@@ -5,7 +5,7 @@
   <artifactId>secuconnect-java-sdk</artifactId>
   <packaging>jar</packaging>
   <name>secuconnect-java-sdk</name>
-  <version>1.8.1</version>
+  <version>1.9.0</version>
   <url>https://github.com/secuconnect/secuconnect-java-sdk</url>
   <description>Secuconnect Java</description>
   <scm>
@@ -72,7 +72,7 @@
       <plugin>
         <groupId>org.apache.maven.plugins</groupId>
         <artifactId>maven-jar-plugin</artifactId>
-        <version>3.1.2</version>
+        <version>3.2.0</version>
         <executions>
           <execution>
             <goals>
@@ -135,7 +135,7 @@
       <plugin>
         <groupId>org.apache.maven.plugins</groupId>
         <artifactId>maven-source-plugin</artifactId>
-        <version>3.1.0</version>
+        <version>3.2.1</version>
         <executions>
           <execution>
             <id>attach-sources</id>
@@ -216,7 +216,7 @@
           <plugin>
             <groupId>org.apache.maven.plugins</groupId>
             <artifactId>maven-shade-plugin</artifactId>
-            <version>3.2.1</version>
+            <version>3.2.2</version>
             <executions>
               <execution>
                 <goals>
@@ -323,13 +323,13 @@
     <java.version>1.7</java.version>
     <maven.compiler.source>${java.version}</maven.compiler.source>
     <maven.compiler.target>${java.version}</maven.compiler.target>
-    <swagger-core-version>1.5.23</swagger-core-version>
+    <swagger-core-version>1.6.0</swagger-core-version>
     <okhttp-version>2.7.5</okhttp-version>
-    <gson-version>2.8.5</gson-version>
-    <jodatime-version>2.10.3</jodatime-version>
+    <gson-version>2.8.6</gson-version>
+    <jodatime-version>2.10.5</jodatime-version>
     <jaxb-api-version>2.3.1</jaxb-api-version>
-    <activemq-stomp-version>5.15.10</activemq-stomp-version>
-    <junit-version>4.12</junit-version>
+    <activemq-stomp-version>5.15.11</activemq-stomp-version>
+    <junit-version>4.13</junit-version>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     <github.global.server>github</github.global.server>
   </properties>
