+++ src/main/java/io/secuconnect/client/ApiClient.java
@@ -186,7 +186,7 @@ public class ApiClient {
     this.lenientDatetimeFormat = true;
 
     // Set default User-Agent.
-    setUserAgent("secuconnect-java-sdk/1.8.1");
+    setUserAgent("secuconnect-java-sdk/1.9.0");
 
     // Setup authentications (key: authentication name, value: authentication).
     authentications = new HashMap<String, Authentication>();
