--- src/main/java/io/secuconnect/client/model/SecupayTransactionProductDTORedirectUrl.java
@@ -1,8 +1,0 @@
-package io.secuconnect.client.model;
-
-/**
- * @deprecated use SecupayRedirectUrl
- */
-public class SecupayTransactionProductDTORedirectUrl extends SecupayRedirectUrl {
-
-}
